finished the task on a Saturday

+++ restaurants/views.py
@@ -59,6 +59,10 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -73,8 +77,13 @@ def restaurant_create(request):
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
+
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+
+    if not request.user.is_staff and request.user != restaurant.owner:
+        return redirect('no-access')
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -82,6 +91,7 @@ def item_create(request, restaurant_id):
             item.restaurant = restaurant
             item.save()
             return redirect('restaurant-detail', restaurant_id)
+    
     context = {
         "form":form,
         "restaurant": restaurant,
@@ -89,7 +99,12 @@ def item_create(request, restaurant_id):
     return render(request, 'item_create.html', context)
 
 def restaurant_update(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+    if not request.user.is_staff and request.user != restaurant_obj.owner:
+        return redirect('no-access')
+    
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +118,13 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+
+    if not request.user.is_staff:
+        return redirect('no-access')
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def no_access(request):
+    return render(request, 'noaccess.html')
